backend/requirements.txt
# JAWIR OS Backend Dependencies
# Python 3.12+ required

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# LangGraph & LangChain
langgraph>=0.0.45
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0

# Web Search & Research
tavily-python>=0.3.0

# Browser Automation (optional, for deep scraping)
playwright>=1.41.0

# Data Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database (for checkpointer)
aiosqlite>=0.19.0

# PDF Processing
PyMuPDF>=1.23.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.0
python-json-logger>=2.0.0
paho-mqtt>=2.0.0  # IoT MQTT integration

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
